+++ src/Entity/StandardFees.php
@@ -25,7 +25,7 @@ class StandardFees
     private $name;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $unitAmount;
 
@@ -56,12 +56,12 @@ class StandardFees
         return $this;
     }
 
-    public function getUnitAmount(): ?int
+    public function getUnitAmount(): ?float
     {
         return $this->unitAmount;
     }
 
-    public function setUnitAmount(int $unitAmount): self
+    public function setUnitAmount(float $unitAmount): self
     {
         $this->unitAmount = $unitAmount;
